requirements.txt
Flask==2.3.3
Flask-Smorest==0.42.1
python-dotenv  
marshmallow
Flask-SQLAlchemy
sqlalchemy
flask-jwt-extended
flask-migrate
passlib
gunicorn
psycopg2
requests
rq
redis
